% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleData.R
\docType{data}
\name{ChoraleAnalyses}
\alias{ChoraleAnalyses}
\title{Numerous harmonic analyses of the 572 chorales.}
\format{A list of length 4 with depth ranging from 1 to 3. All lists
contain (at some level) in a vector or list of length 98,453 (the number of permutational
analyses).}
\usage{
ChoraleAnalyses
}
\description{
Numerous harmonic analyses of the 572 chorales.
}
\keyword{datasets}
