% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{choraleAnalyze}
\alias{choraleAnalyze}
\title{Produce harmonic analyses of chorales}
\usage{
choraleAnalyze(filters, output.dir = "./", output.label = "")
}
\arguments{
\item{filters}{A list of lists. The first element in each list will be used as the
\code{rank} argument to \code{\link{filterAnalyses}}. The second argument
in each list will be used as the \code{filter} argument to \code{\link{filterAnalyses}}.}

\item{output.dir}{A string indicating the path where humdrum files should be output.}

\item{output.label}{A string (defaulting to "") which is appended to the filename of each
outputted file. This is useful for distinguihing analyzes output with separate calls
to \code{choraleAnalyze}.}
}
\value{
A humdrum analysis table, as output be \code{\link{generateHumdrum}} (returned invisibly).
}
\description{
This is the master API for the package. A user may specify
one or more critera for either filtering or ranking analyses. These criteria
will be used to produced specific analyses and output them to new humdrum files.
}
