% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleData.R
\docType{data}
\name{ChoraleTable_Notes}
\alias{ChoraleTable_Notes}
\title{Features parsed from 572 chorale **kern files, by note.}
\format{A data table with 171,915 rows and 40 columns.
\describe{
 \item{Notes}{Tonal letter name of note. "-" = flat. "#" = sharp.}
}}
\usage{
ChoraleTable_Notes
}
\description{
This dataset contains a large set of useful features parsed
from  572 humdrum (**kern) files representing chorales by
J.S. Bach and M. Praetorius, organized by individual note.
}
\details{
The dataset contains 40 pieces of information about each
note position in the data. ("Note position" includes places
where a voice rests or sustains a previously attacked note.)
Twenty-three columns (prefix = "Slice_") pertain to the entire slice---
for instance, the beat position---and are thus duplicated for
notes which occur at the same time. The remaining seventeen variables
actually contain different information for each note.
}
\seealso{
\code{\link{ChoraleTable_Slices}}
}
\keyword{datasets}
