% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleData.R
\docType{data}
\name{ChoraleTable_Slices}
\alias{ChoraleTable_Slices}
\title{Features parsed from 572 chorale **kern files, by slice.}
\format{A data table with 42,928 rows and 40 columns.
\describe{
 \item{Note_Accented}{Tonal letter name of note. "-" = flat. "#" = sharp.}
 \item{Note_TonalName}{Tonal letter name of note. "-" = flat. "#" = sharp.}
}}
\usage{
ChoraleTable_Slices
}
\description{
This dataset contains a large set of useful features parsed
from  572 humdrum (**kern) files representing chorales by
J.S. Bach and M. Praetorius, organized by slice (records
in the original humdrum files).
}
\details{
The dataset contains 40 columns of information about each
slice in the data.Twenty-three columns (prefix = "Slice_") pertain
to the entire slice---for instance, the beat position. The remaining
seventeen variables actually contain different information for each musical
voice in each slice: these are represented as lists of vectors,
with the first element in each vector being the highest voice (e.g., the soprano),
and the last being the lowest (e.g., the bass).
}
\seealso{
\code{\link{ChoraleTable_Notes}}
}
\keyword{datasets}
