% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{filterAnalyses}
\alias{filterAnalyses}
\title{Filter permutational analyses to select specific analyses}
\usage{
filterAnalyses(rank = list(), filter = list())
}
\arguments{
\item{rank}{A list of formulae}

\item{rank}{A list of formulae}
}
\description{
This function selects one analysis for each window in the
\code{\link{PermutationalChoraleAnnotations}} object. It also references
\code{\link{ChoraleTable_Notes}} and \code{\link{ChoraleTable_Slices}} objects.
}
