% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\name{generateHumdrum}
\alias{generateHumdrum}
\title{Align analyses with humdrum data.}
\usage{
generateHumdrum(analyses)
}
\arguments{
\item{analyses}{A list of analyses output by \code{\link{filterAnalyses}}. Each vector
of chord symbols should be the same length as the \code{nrow{\link{ChoraleTable_Slices}}}.}
}
\value{
A list of character veators, with each element of each vector
being corresponding to a single row/slice in the \code{\link{ChoraleTable_Slices}} object.
}
\description{
This function takes one or more harmonic analyses
(output by \code{\link{filterAnalyses}}) and aligns them with the appropriate humdrum
data.
}
