% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{getChordSymbols}
\alias{getChordSymbols}
\title{Extracts vector of chord symols from annotation table}
\usage{
getChordSymbols(DTpermutations)
}
\description{
Extracts vector of chord symols from annotation table
}
