% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{parseFormulae}
\alias{parseFormulae}
\title{Parse formulae used by \code{\link{filterAnalyses}}}
\usage{
parseFormulae(forms)
}
\description{
This function takes a list of formulae and filters and groups them
into three categories: "Windows", "Notes", "Slices." Each formula's
category is determined by the left-side of the formula. If the left side
is empty, the default is "Windows." Left-sides that don't \code{\link{base::pmatch}}
 with one of the three categories are removed.
}
\details{
@param forms A list of formulae.

@return A list (of length three) of lists of expressions.
}
