% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleData.R
\docType{data}
\name{PermutationalChoraleAnnotations}
\alias{PermutationalChoraleAnnotations}
\title{Numerous harmonic analyses of the 572 chorales.}
\format{A data table with 98,453 rows and 12 columns.}
\usage{
PermutationalChoraleAnnotations
}
\description{
This data table contains numerous permutational analyses
of the 572 chorales. The columns \code{Slice_WindowNumber}
and \code{Slice_Number} correspond with the same columns in the
 \code{\link{ChoraleTable_Notes}} and \code{\link{ChoraleTable_Slices}}
objects. The \code{CT_Note_Number} and \code{NCT_Note_Number} columns
contain lists of integers which correspond to the \code{Note_Number} column
in the ChoraleTable_* objects.
}
\keyword{datasets}
