% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{prefer}
\alias{prefer}
\title{Rank preferences}
\usage{
prefer(x, ..., other = max(unlist(list(...))) + 1, combine = max)
}
\arguments{
\item{x}{A list of vectors of rankable values}

\item{...}{Any number of numeric values, each named. Namesless
values are ignored. Each name should correspond to a value in x.}

\item{other}{Any values in x which do not appear as a name in ...
are replaced with other.}
}
\description{
Rank preferences
}
