% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\name{showHumdrum}
\alias{showHumdrum}
\title{Send a humdrum file to the Verovio Humdrum Viewer}
\usage{
showHumdrum(analyses = list(), composer = "Bach", number = 1)
}
\arguments{
\item{analyses}{A list of analyses.}

\item{composer}{A character string: either "Bach" or "Praetorius" (inexact matches are allowed).}

\item{number}{An integer in the range 1--371 (for Bach) or 1--200 (for Praetorious).}
}
\description{
This function sends an annotated humdrum data file (generated by
\code{\link{generateHumdrum}}) to the \href{https://verovio.humdrum.org/}{Verovio Humdrum Viewer}.
}
