% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\name{writeHumdrum}
\alias{writeHumdrum}
\title{Write annotated humdrum data to files}
\usage{
writeHumdrum(fileTable, path = "./", label = NULL)
}
\arguments{
\item{fileTable}{A data table representing annotated humdrum files
(output by \code{\link{generateHudrum}}).}

\item{label}{A character string to append to each filename.
Defaults to NULL. Use this to differentiate different batches
of analyses you generate.}
}
\description{
This function writes new humdrum data (original **kern with added harmonic
analyses), to text files. The original krn file names are used, with the addition of
the word "annotated," a optional label specified by the user, and a different extension (\code{.hum}).
}
